requirements.txt
asn1crypto==0.24.0
boto3==1.9.156
botocore==1.12.156
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.6.1
docutils==0.14
future==0.17.1
idna==2.8
jmespath==0.9.4
lxml==4.3.3
pycparser==2.19
python-dateutil==2.8.0
python-lambda==3.2.6
python-telegram-bot==11.1.0
PyYAML==5.1
requests==2.22.0
s3transfer==0.2.0
six==1.12.0
urllib3==1.25.3
